backend/community-backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
python-dotenv>=1.0
greenlet>=3.0
pydantic>=2.0
pydantic-settings>=2.0
bcrypt>=4.0
PyJWT>=2.10
httpx>=0.28
